Split out savesystem.cpp
