+++ src/gameframework/saveload.cpp
@@ -1,3 +1,7 @@
+// This file has functions from what appear to be several different TUs in the original.
+// Getting a match, especially with optimization levels, may require splitting out functions
+// into new TUs.
+
 #include <stdio.h>
 #include <string.h>
 #include <sys/stat.h>
@@ -7,7 +11,6 @@
 #include "gameframework/saveload.h"
 
 #include "nu2api/nuandroid/ios_graphics.h"
-#include "nu2api/nucore/numemory.h"
 #include "nu2api/nucore/nustring.h"
 #include "nu2api/nucore/nuthread.h"
 #include "nu2api/nucore/nutime.h"
@@ -123,53 +126,6 @@ i32 saveloadLoadSlot(i32 slot, void *buffer, usize size) {
 
 i32 saveloadSaveSlot(i32 slot, void *buffer, usize size) {
     return PCSaveSlot(slot, buffer, static_cast<i32>(size), static_cast<u32>(-1));
-}
-
-typedef i16 (*hashfn_t)(void);
-
-i32 SAVESLOTS = 3;
-
-void *memcard_savedata = NULL;
-i32 memcard_savedatasize = 0;
-void *memcard_savedatabuffer = NULL;
-
-void *memcard_extra_savedata = NULL;
-i32 memcard_extra_savedatasize = 0;
-void *memcard_extra_savedatabuffer = NULL;
-
-hashfn_t memcard_hashfn = NULL;
-void (*memcard_drawasiconfn)(void) = NULL;
-i32 memcard_autosave = 0;
-i32 memcard_autosaveneeded = 0;
-i32 memcard_autosavestarted = 0;
-f32 memcard_autosavepredelay = 0.0f;
-f32 memcard_autosavepostdelay = 0.0f;
-
-void SaveSystemInitialise(i32 slots, void *makeSaveHash, void *save, i32 saveSize, i32 saveCount,
-                          void (*drawSaveIcon)(void), void *extradata, i32 extradataSize) {
-    if (extradata == NULL) {
-        SAVESLOTS = 6;
-        if (slots < 7) {
-            SAVESLOTS = slots;
-        }
-    } else {
-        SAVESLOTS = 5;
-        if (slots < 6) {
-            SAVESLOTS = slots;
-        }
-    }
-
-    memcard_hashfn = (hashfn_t)makeSaveHash;
-    memcard_savedata = save;
-    memcard_savedatasize = saveSize;
-    memcard_extra_savedata = extradata;
-    memcard_extra_savedatasize = extradataSize;
-
-    memcard_savedatabuffer = NU_ALLOC(saveSize + 4, 4, 1, "", NUMEMORY_CATEGORY_NONE);
-    memcard_extra_savedatabuffer = NU_ALLOC(extradataSize + 4, 4, 1, "", NUMEMORY_CATEGORY_NONE);
-
-    memcard_autosave = saveCount;
-    memcard_drawasiconfn = drawSaveIcon;
 }
 
 void createslotfolder(i32 slot) {
